rename.c (simple_copy): Only define if it is going to be used.   (smart_rename): Mark the preserve_dates parameter as possibly	being unused.
resres.c (write_res_data): Prevent a potential compile time warning by casting
  the return value from fwrite.

+++ binutils/resres.c
@@ -408,7 +408,7 @@ write_res_data (data, size, count)
      size_t size;
      int count;
 {
-  if (fwrite (data, size, count, fres) != (size_t) count)
+  if ((size_t) fwrite (data, size, count, fres) != (size_t) count)
     fatal ("%s: could not write to file", filename);
 }
